+++ src/com/koxudaxi/ruff/RuffConfigPanel.kt
@@ -1,20 +1,20 @@
 package com.koxudaxi.ruff
 
 
+import com.intellij.ide.plugins.PluginManagerConfigurable
+import com.intellij.openapi.fileChooser.FileChooserDescriptorFactory
+import com.intellij.openapi.options.ShowSettingsUtil
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.TextFieldWithBrowseButton
-import com.intellij.ui.components.JBTextField
-import com.intellij.openapi.fileChooser.FileChooserDescriptorFactory
 import com.intellij.openapi.ui.emptyText
+import com.intellij.ui.components.JBTextField
 import com.jetbrains.python.sdk.pythonSdk
 import com.koxudaxi.ruff.RuffConfigService.Companion.getInstance
 import org.jetbrains.annotations.SystemDependent
+import java.awt.event.MouseAdapter
+import java.awt.event.MouseEvent
 import java.io.File
-
-import javax.swing.JButton
-import javax.swing.JCheckBox
-import javax.swing.JLabel
-import javax.swing.JPanel
+import javax.swing.*
 
 
 class RuffConfigPanel(project: Project) {
@@ -36,10 +36,17 @@ class RuffConfigPanel(project: Project) {
     private lateinit var ruffConfigPathField: TextFieldWithBrowseButton
     private lateinit var clearRuffConfigPathButton: JButton
     private lateinit var disableOnSaveOutsideOfProjectCheckBox: JCheckBox
-    private lateinit var useRuffLspCheckBox: JCheckBox
+    private lateinit var useRuffLspRadioButton: JRadioButton
+    private lateinit var useIntellijLspClientRadioButton: JRadioButton
+    private lateinit var useLsp4ijRadioButton: JRadioButton
     private lateinit var useRuffFormatCheckBox: JCheckBox
+    private lateinit var useRuffServerRadioButton: JRadioButton
+    private lateinit var enableLspCheckBox: JCheckBox
+    private lateinit var lsp4ijLinkLabel: JLabel
+
     init {
         val ruffConfigService = getInstance(project)
+        val ruffCacheService = RuffCacheService.getInstance(project)
         runRuffOnSaveCheckBox.isSelected = ruffConfigService.runRuffOnSave
         runRuffOnSaveCheckBox.isEnabled = true
         runRuffOnReformatCodeCheckBox.isSelected = ruffConfigService.runRuffOnReformatCode
@@ -49,11 +56,24 @@ class RuffConfigPanel(project: Project) {
         alwaysUseGlobalRuffCheckBox.isEnabled = true
         alwaysUseGlobalRuffCheckBox.isSelected = ruffConfigService.alwaysUseGlobalRuff
         disableOnSaveOutsideOfProjectCheckBox.isEnabled = ruffConfigService.runRuffOnSave
-        useRuffLspCheckBox.isEnabled = lspIsSupported
-        useRuffLspCheckBox.isSelected = ruffConfigService.useRuffLsp
+        useRuffLspRadioButton.isSelected = ruffConfigService.useRuffLsp
+        useRuffLspRadioButton.isEnabled = lspSupported
+        useIntellijLspClientRadioButton.isEnabled = intellijLspClientSupported
+        useIntellijLspClientRadioButton.isSelected = ruffConfigService.useIntellijLspClient
+        useLsp4ijRadioButton.isSelected = ruffConfigService.useLsp4ij
         useRuffFormatCheckBox.isEnabled = true
         useRuffFormatCheckBox.isSelected = ruffConfigService.useRuffFormat
+        useRuffServerRadioButton.isEnabled =  lspSupported
+        useRuffServerRadioButton.isSelected = ruffConfigService.useRuffServer || !ruffConfigService.useRuffLsp
         disableOnSaveOutsideOfProjectCheckBox.isSelected = ruffConfigService.disableOnSaveOutsideOfProject
+        enableLspCheckBox.isEnabled = lspSupported
+        enableLspCheckBox.isSelected = ruffConfigService.enableLsp
+        lsp4ijLinkLabel.addMouseListener(object : MouseAdapter() {
+            override fun mouseClicked(e: MouseEvent?) {
+                 ShowSettingsUtil.getInstance().showSettingsDialog(project, PluginManagerConfigurable::class.java,
+                    { c -> c.openMarketplaceTab("com.redhat.devtools.lsp4ij") })
+            }
+        })
         runRuffOnSaveCheckBox.addActionListener {
             disableOnSaveOutsideOfProjectCheckBox.isEnabled = runRuffOnSaveCheckBox.isSelected
         }
@@ -89,10 +109,22 @@ class RuffConfigPanel(project: Project) {
 
         setAutodetectedRuffLspButton.addActionListener { setAutodetectedRuffLsp() }
 
-        useRuffLspCheckBox.addActionListener { updateLspExecutableFields() }
 
+        useRuffLspRadioButton.addActionListener {
+            updateLspClientCheckBoxes()
+            updateLspExecutableFields()
+        }
+        useRuffServerRadioButton.addActionListener {
+            updateLspClientCheckBoxes()
+            updateLspExecutableFields()
+        }
+
+        enableLspCheckBox.addActionListener {
+            updateLspFields()
+        }
+        updateLspFields()
         alwaysUseGlobalRuffCheckBox.addActionListener { updateProjectExecutableFields() }
-        when (val projectRuffExecutablePath = ruffConfigService.projectRuffExecutablePath?.takeIf { File(it).exists() } ?: getProjectRuffExecutablePath(project, false)) {
+        when (val projectRuffExecutablePath = ruffCacheService.getProjectRuffExecutablePath()?.takeIf { File(it).exists() } ?: getProjectRuffExecutablePath(project, false)) {
             is String -> projectRuffExecutablePathField.text = projectRuffExecutablePath
             else -> {
                 projectRuffExecutablePathField.text = ""
@@ -101,7 +133,7 @@ class RuffConfigPanel(project: Project) {
         }
 
 
-        when (val projectRuffLspExecutablePath = ruffConfigService.projectRuffLspExecutablePath?.takeIf { File(it).exists() } ?: getProjectRuffExecutablePath(project, true)) {
+        when (val projectRuffLspExecutablePath = ruffCacheService.getProjectRuffLspExecutablePath()?.takeIf { File(it).exists() } ?: getProjectRuffExecutablePath(project, true)) {
             is String -> projectRuffLspExecutablePathField.text = projectRuffLspExecutablePath
             else -> {
                 projectRuffLspExecutablePathField.text = ""
@@ -118,9 +150,28 @@ class RuffConfigPanel(project: Project) {
             ruffConfigPathField.text = ""
         }
     }
-
+    private fun updateLspClientCheckBoxes() {
+        val useLspClient = useRuffLspRadioButton.isSelected || useRuffServerRadioButton.isSelected
+        useIntellijLspClientRadioButton.isEnabled = useLspClient && intellijLspClientSupported
+        useLsp4ijRadioButton.isEnabled = useLspClient && lsp4ijSupported
+    }
+    private fun updateLspFields() {
+        if (enableLspCheckBox.isSelected) {
+            useLsp4ijRadioButton.isEnabled = lsp4ijSupported
+            useIntellijLspClientRadioButton.isEnabled =  intellijLspClientSupported
+            useRuffLspRadioButton.isEnabled = lspSupported
+            useRuffServerRadioButton.isEnabled = lspSupported
+        } else {
+            updateLspClientCheckBoxes()
+            updateLspExecutableFields()
+            useLsp4ijRadioButton.isEnabled = false
+            useIntellijLspClientRadioButton.isEnabled = false
+            useRuffLspRadioButton.isEnabled = false
+            useRuffServerRadioButton.isEnabled = false
+        }
+    }
     private fun updateLspExecutableFields() {
-        val enabled = lspIsSupported && useRuffLspCheckBox.isSelected
+        val enabled = intellijLspClientSupported && useRuffLspRadioButton.isSelected
         globalRuffLspExecutablePathField.isEnabled = enabled
         globalRuffLspLabel.isEnabled = enabled
         setAutodetectedRuffLspButton.isEnabled = enabled
@@ -133,7 +184,7 @@ class RuffConfigPanel(project: Project) {
         val enabled = !alwaysUseGlobalRuffCheckBox.isSelected
         projectRuffExecutablePathField.isEnabled = enabled
         projectRuffLabel.isEnabled = enabled
-        if (lspIsSupported && useRuffLspCheckBox.isSelected) {
+        if (intellijLspClientSupported && useRuffLspRadioButton.isSelected) {
             projectRuffLspExecutablePathField.isEnabled = enabled
             projectRuffLspLabel.isEnabled = enabled
         }
@@ -177,10 +228,17 @@ class RuffConfigPanel(project: Project) {
     val disableOnSaveOutsideOfProject: Boolean
         get() = disableOnSaveOutsideOfProjectCheckBox.isSelected
     val useRuffLsp: Boolean
-        get() = useRuffLspCheckBox.isSelected
-
+        get() = useRuffLspRadioButton.isSelected
+    val useIntellijLspClient: Boolean
+        get() = useIntellijLspClientRadioButton.isSelected
+    val useLsp4ij: Boolean
+        get() = useLsp4ijRadioButton.isSelected
+    val useRuffServer: Boolean
+        get() = useRuffServerRadioButton.isSelected
     val useRuffFormat: Boolean
         get() = useRuffFormatCheckBox.isSelected
+    val enableLsp: Boolean
+        get() = enableLspCheckBox.isSelected
 
     companion object {
         const val RUFF_EXECUTABLE_NOT_FOUND =  "Ruff executable not found"