+++ src/com/koxudaxi/ruff/RuffProjectInitializer.kt
@@ -1,6 +1,5 @@
 package com.koxudaxi.ruff
 
-import RuffLspServerSupportProvider
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.DumbService
 import com.intellij.openapi.project.Project
@@ -9,7 +8,6 @@ import com.intellij.openapi.startup.ProjectActivity
 import com.intellij.openapi.vfs.AsyncFileListener
 import com.intellij.openapi.vfs.VirtualFileManager
 import com.intellij.openapi.vfs.newvfs.events.*
-import com.intellij.platform.lsp.api.LspServerManager
 import com.intellij.serviceContainer.AlreadyDisposedException
 
 class RuffProjectInitializer : ProjectActivity {
@@ -21,9 +19,9 @@ class RuffProjectInitializer : ProjectActivity {
             try {
                 val ruffCacheService = RuffCacheService.getInstance(project)
                 if (ruffCacheService.getVersion() == null) {
-                    ruffCacheService.setVersion()
+                    ruffCacheService.setVersion{}
                 }
-                if (lspIsSupported) {
+                if (lspSupported) {
                     setUpPyProjectTomlLister(project)
                 }
             } catch (_: AlreadyDisposedException) {
@@ -33,8 +31,7 @@ class RuffProjectInitializer : ProjectActivity {
 
     private fun setUpPyProjectTomlLister(project: Project) {
         val ruffConfigService = RuffConfigService.getInstance(project)
-        @Suppress("UnstableApiUsage")
-        val lspServerManager = LspServerManager.getInstance(project)
+        val ruffLspClientManager = RuffLspClientManager.getInstance(project)
 
         VirtualFileManager.getInstance().addAsyncFileListener(
             { events ->
@@ -56,11 +53,12 @@ class RuffProjectInitializer : ProjectActivity {
                             ) return
                             ApplicationManager.getApplication().invokeLater {
                                 if (project.isDisposed) return@invokeLater
-                                if (!ruffConfigService.useRuffLsp) return@invokeLater
-                                lspServerManager.stopAndRestartIfNeeded(RuffLspServerSupportProvider::class.java)
-
-                                LspServerManager.getInstance(project)
-                                    .stopAndRestartIfNeeded(RuffLspServerSupportProvider::class.java)
+                                if (!ruffLspClientManager.hasClient()) return@invokeLater
+                                if (!ruffConfigService.enableLsp) {
+                                    ruffLspClientManager.stop()
+                                } else {
+                                    ruffLspClientManager.restart()
+                                }
                             }
                         } catch (_: AlreadyDisposedException) {
                         }